requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
aiohttp==3.8.4
pytest==7.3.1
httpx==0.23.3
PyGithub==1.58.1
python-multipart==0.0.6
pyngrok==6.0.0
